requirements.txt
Flask>=2.0.0,<3.0.0
Flask-CORS>=3.0.0
mysql-connector-python>=8.0.0
gpiozero>=1.6.0
python-dotenv>=0.19.0
numpy>=1.19.5
pytz>=2021.1
typing-extensions>=3.7.4
pytest>=6.2.0
RPi.GPIO>=0.7.0
pymssql>=2.0.0